Largen byte size of totalbyes
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@12322 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ daq/rawdata/modules/include/Serializer.h
@@ -177,7 +177,7 @@ namespace Belle2 {
     CprErrorMessage print_err;
 
     //! sent data size
-    int m_totbytes;
+    unsigned long long m_totbytes;
 
     //! start flag
     int m_start_flag;
